+++ agent/src/main/java/com/teamscale/jacoco/agent/options/AgentOptions.java
@@ -46,6 +46,7 @@ import java.io.IOException;
 import java.lang.instrument.Instrumentation;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -344,9 +345,11 @@ public class AgentOptions {
 	}
 
 	private void validateTestwiseCoverageConfig(Validator validator) {
+		boolean diskMode = testImpactConfig.testwiseCoverageMode == ETestwiseCoverageMode.DISK;
+
 		validator.isFalse(
-				httpServerPort == null && testImpactConfig.testEnvironmentVariable == null,
-				"You use 'mode' 'TESTWISE' but did use neither 'http-server-port' nor 'test-env'!" +
+				!diskMode && httpServerPort == null && testImpactConfig.testEnvironmentVariable == null,
+				"You use 'mode' 'TESTWISE' but did use neither 'http-server-port', 'test-env', nor dumping to disk!" +
 						" One of them is required!");
 
 		validator.isFalse(
@@ -512,6 +515,30 @@ public class AgentOptions {
 		org.conqat.lib.commons.filesystem.FileSystemUtils.ensureDirectoryExists(outputDirectory.toFile());
 		return outputDirectory.resolve(prefix + "-" + LocalDateTime.now().format(DATE_TIME_FORMATTER) + "." + extension)
 				.toFile();
+	}
+
+	/**
+	 * Creates a new file with the given prefix, extension and current timestamp and ensures that the parent folder
+	 * actually exists. One output folder is created per partition.
+	 */
+	public File createNewFileInPartitionOutputDirectory(String prefix, String extension) throws IOException {
+		Path partitionOutputDirectory = Paths.get(outputDirectory.toString(), safeFolderName(getTeamscaleServerOptions().partition));
+		org.conqat.lib.commons.filesystem.FileSystemUtils.ensureDirectoryExists(partitionOutputDirectory.toFile());
+		return partitionOutputDirectory.resolve(prefix + "-" + LocalDateTime.now().format(DATE_TIME_FORMATTER) + "." + extension)
+				.toFile();
+	}
+
+	private static String safeFolderName(String folderName) {
+		String result = folderName.replaceAll("[<>:\"/\\|?*]", "")
+				.replaceAll("\\.{1,}", "dot")
+				.replaceAll("\\x00", "")
+				.replaceAll("[. ]$", "");
+
+		if (result.isEmpty()) {
+			return "default";
+		} else {
+			return result;
+		}
 	}
 
 	/**