+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/CoverageToDiskStrategy.java
@@ -12,16 +12,16 @@ import java.io.IOException;
 /**
  * Strategy for appending coverage into one json test-wise coverage file with one session per test.
  */
-public class CoverageToJsonFileStrategy extends CoverageToJsonStrategyBase {
+public class CoverageToDiskStrategy extends CoverageToJsonStrategyBase {
 
-	public CoverageToJsonFileStrategy(JacocoRuntimeController controller, AgentOptions agentOptions,
-									  JaCoCoTestwiseReportGenerator reportGenerator) {
+	public CoverageToDiskStrategy(JacocoRuntimeController controller, AgentOptions agentOptions,
+								  JaCoCoTestwiseReportGenerator reportGenerator) {
 		super(controller, agentOptions, reportGenerator);
 	}
 
 	@Override
 	protected void handleTestwiseCoverageJsonReady(String json) throws IOException {
-		File reportFile = agentOptions.createNewFileInOutputDirectory("testwise-coverage", "json");
+		File reportFile = agentOptions.createNewFileInPartitionOutputDirectory("testwise-coverage", "json");
 		FileSystemUtils.writeFileUTF8(reportFile, json);
 	}
 }