+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageAgent.java
@@ -28,8 +28,8 @@ public class TestwiseCoverageAgent extends AgentBase {
 			case TEAMSCALE_UPLOAD:
 				testEventHandler = new CoverageToTeamscaleStrategy(controller, options, reportGenerator);
 				break;
-			case JSON_FILE:
-				testEventHandler = new CoverageToJsonFileStrategy(controller, options, reportGenerator);
+			case DISK:
+				testEventHandler = new CoverageToDiskStrategy(controller, options, reportGenerator);
 				break;
 			case HTTP:
 				testEventHandler = new CoverageViaHttpStrategy(controller, options, reportGenerator);